1147: Initial support of Broadcom VideoCore IV
VideoCore IV: Simplify signext instruction

The conditional jump is difficult to handle in the decompiler. A
left shift followed by a (signed) right shift is easier.

VideoCore IV: Avoid an unnecessary COPY in loop instructions

When comparing to an immediate, a copy must be made. Currently, the
value is copied twice: once in loophckcond to check the condition
and once to display the value in the disassembly. However, the
disassembly does not need the value itself; it can use a display
stub and save one unnecessary P-Code.

VideoCore IV: Add a __noclobber function prototype

Some functions in the firmware seem to guarantee that no registers
are clobbered except r0. The caller then uses original r1-r5 values
after the function call.

VideoCore IV: Fix rounding instructions

FLOAT_FLOOR does not produce a twos complement integer output, so:

  1. ffloor and fceil do not need an int2float()
  2. floor needs a trunc()