3137: Include cspec docs in build
Use canonical URL for docbook stylesheets.

Thanks to the magic of XML catalogs, this should not involve
actual network access when the stylesheets are properly
installed.

Squashed:

Ensure that xsltproc doesn't hit the net

The --nonet flag works nicely for this.

Include cspec docs in build

This is a rework of https://github.com/NationalSecurityAgency/ghidra/pull/2108 that does less.

* Mostly copying the layout of the sleigh docs
* We already adjusted the references to the docbook-xsl stylesheets to use the canonical URL,
  so it already works fine with Debian's docbook-xsl package

I didn't commit the generated html into GhidraDocs/ directory, as my docbook-xsl is a slightly
different version that generates slightly different output.

Co-Authored-By: R. Elliott Childre <elliottchildre329@gmail.com>
Signed-off-by: Samuel Bronson <naesten@gmail.com>

Tell gradle that help/docs depend on their sources.

Rather than listing the exact sources needed for each one,
which would be laborious and error prone, just mark all of these
as depending on the whole src/main/doc directory.

Signed-off-by: Samuel Bronson <naesten@gmail.com>

Add missing '/' to path prefix in decompileplugin_pdf.xsl

Apparently this value just gets prepended as-is,
so a trailing slash is kind of important here.

Signed-off-by: Samuel Bronson <naesten@gmail.com>