3534: correct export file validation
Program export validation occurred on the value of the text input box
rather than the effective export file, which could cause false pass
and fails in the validation itself. This change corrects the validation
to occur on the correct file prior to export.

Several tests for the exporter plugin are also included, to test
expected functionality and prevent future regression.