3741: Set isValid to false if Character.isJavaIdentifierPart(ch) is false
This looks like a mistake. The reason why it does is because the loop will only run if it is valid, and once isValid is true, |= means it will never be false, when it runs, meaning the check does nothing.

Instead, let's iterate over the string and if a character is invalid, the entire string is not an identifier