4103: WebAssembly support.
Merged from https://github.com/nneonneo/ghidra-wasm-plugin/.

Closes #2937.

WebAssembly: add a help page to explain the finer points of the extension.

I was hoping to add this to a Processors help submenu, but it doesn't exist,
so I put it in the Appendix instead for now.

WebAssembly: make funcref/externref 32 bits in size.

The use of 64 bits was mostly a debugging feature to ensure we didn't accidentally
mix references with normal types. 64 bit addresses cause some problems, though,
such as 64-bit immediates (e.g. from ref.func) not always being treated as addresses
in the decompiler.

WebAssembly: reduce dependency on Program class.

Using Program throughout is rather lazy, and in many cases totally unnecessary.
Using finer-grained interfaces and classes like AddressFactory allows usage of
the code in non-Program contexts, such as emulation.

WebAssembly: load .module directly to 0x8000000 in RAM

This allows the whole module to be loaded in memory and eliminates the previous
duplication of the .code bytes in the .module and the .function bytes in RAM.

WebAssembly: extend regs to 16 bytes to prepare for vector insts

Vector instructions will require 128-bit values, so prepare for that change by
widening all registers to 128 bits.

WebAssembly: Emulation support

This patch implements a EmulateInstructionStateModifier for Wasm which provides
support for emulating Wasm instructions.

The memory contents must contain a full module for this to work, because
instruction semantics still depend on module details (e.g. the type of certain
operations depends on metadata like the types of imports or globals).

WebAssembly: correctly handle u32 opcodes.

0xFC (and 0xFD) have a u32 as the second operand, so we need to handle this
correctly.

WebAssembly: SIMD support.

Disassembly and verification are implemented, but almost all of the SIMD opcodes
are just stubbed out with pcodeops for now, so semantics aren't implemented. This
is probably good enough for now.

WebAssembly: create code length data to avoid undefined bytes

WebAssembly: update help stylesheet to recommended style

WebAssembly: remove now-deleted DataTypeConflictException

WebAssembly: update to use standard LEB128 utilities

WebAssembly: disable 2+ output vars for pos-stack as it is unsupported.

Without this, we get the error "<pentry> tags within a group must be
distinguished by size or type" when attempting to load the pos-stack compiler
spec. This will break functions that have multiple output arguments, but those
are expected to be rare (and they are not well-supported by Ghidra anyway).

WebAssembly: Add a new script to handle "flat" dyncall namespaces.

It seems like LLVM, for instance, uses table indices as function pointers, so
this script is likely to be useful for anything compiled with LLVM. As a guess,
analyze_dyncalls is probably only useful for programs compiled using the
Emscripten fastcomp backend.

WebAssembly: Mark __wasm prototypes as register strategies

WebAssembly: Add a stack analyzer which creates C stack variables for Wasm.

The traditional stack analyzer, StackVariableAnalyzer
(NewFunctionStackAnalysisCmd) depends on having registers that contain stack
addresses visible as assembly operands. However, due to the Wasm disassembler
design, which hides the Wasm stack registers from the disassembly, this stack
analysis cannot automatically extract stack variables.

As a fix, adapt FunctionStackAnalysisCmd to operate on the hidden Wasm stack
register operands to detect operations that interact with C stack addresses.

This helps with cross references to C stack variables, as well as allowing C
stack variables to be properly interacted with in the decompiler.

WebAssembly: update ConstantExpression comment

WebAssembly: add wasm build instructions to pcodetest

This successfully builds a working binary with Clang 15.

- `mem*` definitions in `misc_BODY.c` and new `main` definition in `tpp.py`
  are for C99 compatibility
- `encoding='utf8'` in build.py produces a more readable log
- `TestInfo_force` in pcode_test.c ensures that the entire `MainInfo` structure
  is included in the binary

WebAssembly: implement pcodetest tests.

We pass almost every test with -O0 and -O3, with the exception of pcode_conversions_Main.

There's a rather ugly hack needed to convert certain function pointers from
table indices (used by the actual code) to byte addresses (used by the emulator
to set PC). The way this is implemented is decidedly not ideal; a much better
solution would be to somehow hook readCodePointer in the
ProcessorEmulatorTestAdapter subclass; this would also enable the "procedure
descriptor indirection" fix to be moved into a processor-specific
implementation.

Changing the structs to protected in `ProcessorEmulatorTestAdapter` is for
convenience, so that the subclass does not need to go look those up again.

The zero-size check added to `BytesPcodeExecutorStateSpace#read` fixes a bug
which caused an exception when reading 0 bytes (`offset + size - 1` is not a
valid calculation in that case).

WebAssembly: Fix pcode_conversions_Main test.

This is fixed by implementing unsigned-to-float conversions properly.

WebAssembly: add a SIMD pcodetest.

Similarly to the x86 AVX2 test, this will allow us to do (basic) testing of Wasm
SIMD opcodes via autovectorization. Only the -O3 binary is expected to contain
SIMD opcodes.

WebAssembly: Simplify laneidx vector operands.

laneidx is actually just a byte, not a LEB128. Attempting to use 16 arguments in
a pcodeop hits a hardcoded limit of 8 args in PcodeEmit, so switch i8x16.shuffle
to using 4 32-bit arguments instead. This is the same syntax as used by
wasm-objdump by default.

WebAssembly: fix load32/load64, simplify some SIMD pcode

load32/load64 were incorrectly taking a laneidx, causing disassembly failures.
Also, take this opportunity to implement some of the SIMD operations using
common pcodeops. This will reduce emulation burden, as well as making loads and
stores explicit for dataflow analysis.