5004: Feature: UNIX A.out Loader
Address review comments; fix bss block init

This addresses review comments concerning (a) the unreliability of A.out format
consistency for OLDSUN2 files, (b) the behavior of NetBSD with respect to header
field endianness, and (c) formatting errors in some of the language specs.

This also adds generation of the .bss block when the object file specifies a
nonzero size for .bss.